See details in commit message body
Prefixed all commands with a dot; that way it less likely for a command to be confused with a regulat statement. E.g `next` is now `.next`. The only commands not changed are `!` and `!!` and `!-`

Also removed unnecessary colon form `add` and `rm` commands

Used `describe` action to show question queue state

Used methods to add and remove admin from admin list

Bug fixes and minor changes