Updated README with more detailed info.
Modified configuration.

Modified message processing such that, it is clear which messages
should be sent in PM versus which ones should be sent in channel;
which messages should come from admin versus any user.
This is also specified in the README.

Move fpaste call to "logger" module. This call happens when a class
session ends, and the returned URL is stored for later access by any user
who issues the command `.givemelogs`.

Changed name of `QuestionQueue` class to `Queue`.